+++ src/modules/DisableSaveConfirmation.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "module/Module.hpp"
+
+class $editorModule(DisableSaveConfirmation) {
+    bool m_setSaveHijack = false;
+
+    bool onToggled(bool state) override;
+    void onEditor() override;
+
+    void setHijack(EditorPauseLayer* pauseLayer);
+};