+++ src/modules/ImprovedColorPicker.cpp
@@ -6,6 +6,7 @@
 #include "utils/NextFree/NextFreeOffsetInput.hpp"
 
 bool ImprovedColorPicker::onToggled(bool state) {
+    m_toggledHooks.toggle(state);
     return true;
 }